Refactor: Rename command files to match command names, clean up bin types
File renames (preserving git history via git mv):
- channel-videos.ts  → list-videos.ts   (command: list-videos)
- video-info.ts      → get-video.ts     (command: get-video / get-videos)
- update-metadata.ts → update-video.ts  (command: update-video)
- search-channel.ts  → search-videos.ts (command: search-videos)

bin/staqan-yt.ts cleanup:
- Update import paths and variable names to match new filenames
- Replace 2 inline option types with imported ChannelAnalyticsOptions
  and OutputOption & VerboseOption from types/index.ts
- Add types/index.ts import
